+++ infrastructure/tooling/src/dev/postgres.js
@@ -56,44 +56,74 @@ const postgresPrompts = ({ userConfig, prompts, configTmpl }) => {
 };
 
 const postgresResources = async ({ userConfig, answer, configTmpl }) => {
-  let servicesNeedingConfig = [];
+  let servicesNeedingUrls = [];
   for (const [name, cfg] of Object.entries(configTmpl)) {
+    // only examine services in configTmpl..
+    if (!cfg.read_db_url) {
+      continue;
+    }
+
     if (!userConfig[name]) {
       userConfig[name] = {};
     }
-    if (cfg.db_crypto_keys !== undefined && !userConfig[name].db_crypto_keys) {
-      userConfig[name].db_crypto_keys = [{
-        id: 'dev-init',
-        algo: 'aes-256',
-        key: crypto.randomBytes(32).toString('base64'),
-      }];
+
+    // set up crypto keys, including migrating from azure_{crypto,signing}_key
+
+    if (cfg.db_crypto_keys !== undefined) {
+      if (!userConfig[name].db_crypto_keys) {
+        userConfig[name].db_crypto_keys = [{
+          id: 'dev-init',
+          algo: 'aes-256',
+          key: crypto.randomBytes(32).toString('base64'),
+        }];
+      }
+
+      if (userConfig[name].azure_crypto_key) {
+        userConfig[name].db_crypto_keys.unshift({
+          id: 'azure',
+          algo: 'aes-256',
+          key: userConfig[name].azure_crypto_key,
+        });
+        delete userConfig[name].azure_crypto_key;
+      }
+    } else {
+      // this service doesn't use encrypted columns..
+      delete userConfig[name].azure_crypto_key;
     }
+
+    // signing is no longer supported in any service
+    delete userConfig[name].azure_signing_key;
+
+    // check for the *_db_url parameters
+
     if (cfg.read_db_url !== undefined && !userConfig[name].read_db_url) {
-      servicesNeedingConfig.push(name);
+      servicesNeedingUrls.push(name);
     } else if (cfg.write_db_url !== undefined && !userConfig[name].write_db_url) {
-      servicesNeedingConfig.push(name);
+      servicesNeedingUrls.push(name);
     }
   }
 
   // if all services are set up, there's nothing to do..
-  if (servicesNeedingConfig.length === 0) {
+  if (servicesNeedingUrls.length === 0) {
     return userConfig;
   }
 
   const { dbAdminUsername, dbAdminPassword, dbName, dbPublicIp, dbPrivateIp } =
     Object.assign({}, userConfig.meta || {}, answer.meta || {});
-  const adminDbUrl = makePgUrl({
-    hostname: dbPublicIp,
-    username: dbAdminUsername,
-    password: dbAdminPassword,
-    dbname: dbName,
-  });
 
-  const client = new Client({ connectionString: adminDbUrl });
+  const client = new Client({
+    host: dbPublicIp,
+    user: dbAdminUsername,
+    password: dbAdminPassword,
+    database: dbName,
+    // Cloud SQL servers support TLS but not cert validation, so don't
+    // try to validate it.
+    ssl: { rejectUnauthorized: false },
+  });
   await client.connect();
 
   try {
-    for (let serviceName of servicesNeedingConfig) {
+    for (let serviceName of servicesNeedingUrls) {
       const username = `${dbAdminUsername}_${serviceName}`;
       const password = `${slugid.v4()}${slugid.v4()}`;
       const url = makePgUrl({