+++ services/github/test/api_test.js
@@ -253,7 +253,7 @@ helper.secrets.mockSuite(testing.suiteName(), [], function(mock, skipping) {
     await testing.fakeauth.withAnonymousScopes(['github:latest-status:abc123:*'], async () => {
       await assert.rejects(() => got(
         helper.apiClient.buildUrl(helper.apiClient.latest, 'abc123', 'unknownRepo', 'nosuch'),
-        { followRedirect: false }), err => err.statusCode === 404);
+        { followRedirect: false }), err => err.response.statusCode === 404);
     });
   });